requirements.txt
streamlit
sentence-transformers
fastapi
uvicorn
faiss-cpu
pandas
numpy
requests